Release version of Swedish Radio Programs app
Added stretch goal level.
ListView widget replaced with RecyclerView.
This enables an automatic management
of recycling views that go out of display range.
The approach here is to load the entire Swedish Radio
Open API program information data set into the Android device memory
(currently 744 program pages). These exclude the detail data pages,
and limit the app to operate on 40 entries at a time.
This proved to work well.

The app shows a list of programs.
There is a scrollbar that makes it possible to reach
the 744 entries a bit quicker that the scroll method.
Clicking a list entry brings up the program detail
page in a WebView.  This proved to work well
and enables audio playback of program content.

No Unit test code was developed.
Tried to clean up code as best possible in limited time
and used the Android Studio Code analysis tool to find issues.

/gunnar